+++ python/fastdb4py/__init__.py
@@ -10,10 +10,11 @@ from .column_engine import ColumnEngine
 from .object_engine import ObjectEngine
 from .orm.table import Table
 from .serializer import FastSerializer
+from .string_column import StringColumn
 
 __all__ = [
     'feature', 'is_feature', 'get_schema', 'lookup_class',
-    'Layout', 'ColumnEngine', 'ObjectEngine', 'Table',
+    'Layout', 'ColumnEngine', 'ObjectEngine', 'Table', 'StringColumn',
     'FastSerializer',
     'BOOL', 'U8', 'U16', 'U32', 'I32', 'U8N', 'U16N',
     'F32', 'F64', 'STR', 'WSTR', 'REF', 'BYTES',